Rename, removed redundant code from runner
